Move app config to shared package
