+++ package.json
@@ -9,6 +9,7 @@
   "homepage": "https://RoundingWell.com",
   "type": "module",
   "workspaces": [
+    "packages/care-ops-config",
     "packages/care-ops-datadog",
     "packages/care-ops-fontawesome"
   ],
@@ -86,6 +87,7 @@
     "@auth0/auth0-spa-js": "^2.1.3",
     "@fortawesome/fontawesome-svg-core": "^6.1.0",
     "@kinde-oss/kinde-auth-pkce-js": "^4.2.0",
+    "@roundingwell/care-ops-config": "workspace:*",
     "@roundingwell/care-ops-datadog": "workspace:*",
     "@workos-inc/authkit-js": "^0.12.0",
     "amazon-quicksight-embedding-sdk": "^2.10.0",
