+++ src/js/auth/e2e.js
@@ -1,5 +1,5 @@
 import { isEmpty } from 'underscore';
-import { e2eConfig as config } from 'js/config';
+import { e2eConfig as config } from '@roundingwell/care-ops-config';
 import { PATH_ROOT, PATH_LOGOUT } from './config';
 
 let token = config.token;
