+++ src/js/auth/kinde.js
@@ -1,7 +1,7 @@
 import { extend, isEmpty } from 'underscore';
 import createKindeClient from '@kinde-oss/kinde-auth-pkce-js';
 
-import { kindeConfig as config, appConfig } from 'js/config';
+import { kindeConfig as config, appConfig } from '@roundingwell/care-ops-config';
 
 import { LoginPromptView } from 'js/views/globals/prelogin/prelogin_views';
