+++ src/js/entities-service/clinicians.js
@@ -1,4 +1,4 @@
-import { setUser } from 'js/datadog';
+import { setUser, startRum } from 'js/datadog';
 import { v4 as uuid } from 'uuid';
 import Radio from 'backbone.radio';
 import BaseEntity from 'js/base/entity-service';
@@ -18,6 +18,7 @@ const Entity = BaseEntity.extend({
     return this.fetchBy('/api/clinicians/me')
       .then(currentUser => {
         setUser(currentUser.pick('id', 'name', 'email'));
+        startRum();
         currentUser.clientKey = uuid();
         return currentUser;
       });
